+++ src/misc.h
@@ -3,7 +3,7 @@
 
 #include <RcppArmadillo.h>
 
-int factorial(int);
+long int factorial(int);
 int binomial_coefficient(int, int);
 arma::uvec std_setdiff(arma::uvec&, arma::uvec&);
 int sample_int(const arma::rowvec& probs);
